Extract test fixtures
